Access Token and Refresh Token

+++ src/utils/cloudinary.js
@@ -1,32 +1,40 @@
 import { v2 as cloudinary } from 'cloudinary';
-import fs from 'fs'
+import fs from 'fs';
+import { configDotenv } from 'dotenv';
 
-(async function() {
+configDotenv('./env');
 
+;(async function() {
     // Configuration
-    cloudinary.config({ 
-        cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-        api_key: process.env.CLOUDINARY_API_KEY, 
-        api_secret: process.env.CLOUDINARY_API_SECRET 
-    })
-})()
+    cloudinary.config({
+        cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+        api_key: process.env.CLOUDINARY_API_KEY,
+        api_secret: process.env.CLOUDINARY_API_SECRET,
+    });
+})();
 
-const uploadOnCloudinary = async (localFilePath)=>{
-
+const uploadOnCloudinary = async (localFilePath) => {
     try {
-        if(!localFilePath){
-            return null
-        }else{
-          const response =  cloudinary.uploader.upload(localFilePath, {
-                resource_type: 'auto'
-            })
-            console.log('File has been uploaded to cloudinary', (await response).url)
-            return response
-        } 
-    } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved file if upload gets failed 
-        return null
-    }
-}
+        if (!localFilePath) {
+            return null;
+        }
 
-export {uploadOnCloudinary}
+        const response = await cloudinary.uploader.upload(localFilePath, {
+            resource_type: 'auto'
+        });
+
+        console.log('File has been uploaded to cloudinary :: url :', response.url);
+        fs.unlinkSync(localFilePath)
+        
+        return response;
+
+    } catch (error) {
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+        // console.log("line 30 error:: ", error);
+        return null;
+    }
+};
+
+export { uploadOnCloudinary };
